Move test_logging to a separate file.
