+++ tests/test_cli/test_cli.py
@@ -56,14 +56,3 @@ def test_dash_as_name(python_version):
         path_output, "videos", "-", "480p15", "DashAsNameTest.mp4")), err
     rmtree(path_output)
 
-def test_logging_to_file(python_version):
-    """Test logging Terminal output to a log file."""
-    path_basic_scene = os.path.join("tests", "tests_data", "basic_scenes.py")
-    path_output = os.path.join("tests_cache", "media_temp")
-    command = [python_version, "-m", "manim", path_basic_scene,
-               "SquareToCircle", "-l", "--log_to_file", "--log_dir",os.path.join(path_output,"logs"), "--media_dir", path_output]
-    out, err, exitcode = capture(command)
-    assert exitcode == 0, err
-    assert os.path.exists(os.path.join(
-        path_output, "logs", "SquareToCircle.log")), err
-    rmtree(path_output)
